recent changes
Here are my changes I made so far. Christian if you can look at it to see how to print the recognized words on the screen by using the string recognizedWords that you created. Right now you have it sprinting the speechService.speech.lastRecognizedWords and that is only picking up one or two words.